Make list smaller
