--- docs/dev-guide/index.md
@@ -1,74 +1,0 @@
----
-id: index
-title: Temporal developer's guide
-description: The Temporal developer's guide provides a comprehensive overview of the structures, primitives, and features used in Temporal Application development.
-sidebar_label: Dev guide
-sidebar_position: 3
----
-
-:::info WORK IN PROGRESS
-
-This guide is a work in progress.
-Some sections may be incomplete or missing for some languages.
-Information may change at any time.
-
-If you can't find what you are looking for in the Go, Java, TypeScript, PHP Developer's guide, it could be in [older docs for SDKs](https://legacy-documentation-sdks.temporal.io/).
-
-:::
-
-The Temporal SDK developer guides are meant to provide a comprehensive overview of the structures, primitives, and features used in [Temporal Application](/temporal#temporal-application) development.
-
-## Supported SDKs
-
-### Go
-
-Visit the [Go SDK dev guide](/dev-guide/go)!
-
-<div class="app-dev-landing-page-lang-logo">
-<a href="/dev-guide/go">
-<img src="/app-dev/go-lang.svg" alt="Go lang logo" />
-</a>
-</div>
-
-### Java
-
-Visit the [Java SDK dev guide](/dev-guide/java)!
-
-<div class="app-dev-landing-page-lang-logo">
-<a href="/dev-guide/java">
-<img src="/app-dev/java.svg" alt="Java logo" />
-</a>
-</div>
-
-### TypeScript
-
-Visit the [TypeScript SDK dev guide](/dev-guide/typescript)!
-
-<div class="app-dev-landing-page-lang-logos">
-<a href="/dev-guide/typescript?lang=ts">
-<img src="/app-dev/typescript.svg" alt="TypeScript logo" />
-</a>
-<a href="/dev-guide/typescript?lang=js">
-<img src="/img/javascript.svg" alt="JavaScript logo" />
-</a>
-</div>
-
-### Python
-
-Visit the [Python SDK dev guide](/dev-guide/python)!
-
-<div class="app-dev-landing-page-lang-logo">
-<a href="/dev-guide/python">
-<img src="/app-dev/python.svg" alt="python logo" />
-</a>
-</div>
-
-### PHP
-
-Visit the [PHP SDK dev guide](/dev-guide/php)!
-
-<div class="app-dev-landing-page-lang-logo">
-<a href="/dev-guide/php">
-<img src="/app-dev/php.svg" alt="PHP logo" />
-</a>
-</div>
