+++ docs/dev-guide/index.mdx
@@ -1,0 +1,36 @@
+---
+id: index
+title: Temporal developer's guide
+description: The Temporal developer's guide provides a comprehensive overview of the structures, primitives, and features used in Temporal Application development.
+sidebar_label: Dev guide
+sidebar_position: 3
+---
+
+import { SdkLogos } from "../components/SdkLogos";
+
+:::info WORK IN PROGRESS
+
+This guide is a work in progress.
+Some sections may be incomplete or missing for some languages.
+Information may change at any time.
+
+If you can't find what you are looking for in the Go, Java, TypeScript, PHP Developer's guide, it could be in [older docs for SdkLogos](https://legacy-documentation-sdks.temporal.io/).
+
+:::
+
+The Temporal SDK developer guides are meant to provide a comprehensive overview of the structures, primitives, and features used in [Temporal Application](/temporal#temporal-application) development.
+
+## Supported SDKs
+
+<SdkLogos />
+
+- [Go SDK dev guide](/dev-guide/go)
+- [Java SDK dev guide](/dev-guide/java)
+- [PHP SDK dev guide](/dev-guide/php)
+- [Python SDK dev guide](/dev-guide/python)
+- [TypeScript SDK dev guide](/dev-guide/typescript)
+
+The following SDKs do not yet have dev guides:
+
+- .NET SDK: [docs](https://github.com/temporalio/sdk-dotnet#readme) and [API reference](https://dotnet.temporal.io/)
+- Ruby SDK: [docs](https://github.com/temporalio/sdk-ruby#readme)
